Implement extras/Coffee Passport
Build out the inital design for the Coffee Passport (CP) section
in `extras`.

Changes the logo color to blue to match clickable links and to free
up green for use with the "Decaf: Available!" attribute on
Coffee Passport entries. This way, the logo and the decaf attribute
don't match in color, implying they're both clickable.

Also, add the star rating for CP entries to the list view on the
main CP index. This does not conflict with regular list items on
the rest of the site.